requirement.txt
# Core
python-dotenv>=1.0.1
requests>=2.32.3
urllib3>=2.2.3

# Database
psycopg2-binary>=2.9.11
pgvector>=0.2.5

# Embedding + LLM
sentence-transformers>=3.0.0
transformers>=4.44.0
torch>=2.3.1

# OpenAI / Azure / API clients
openai>=1.30.1
tiktoken>=0.7.0

# Utilities
pandas>=2.2.2
numpy>=1.26.4

# Testing
pytest>=8.4.2

# Frontend / Dashboard
streamlit>=1.50.0

# Optional but useful
rich>=13.7.1        # สวยงามเวลา log/debug
